extensions/rally/api-rally/pyproject.toml
[tool.poetry]
name = "api-rally"
version = "0.1.0"
description = "Rally Tascas Service"
authors = []

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^2.0.22"
pydantic = "^2.4.2"
loguru = "^0.7.2"
python-multipart = "^0.0.6"
fastapi = "^0.109.1"
python-jose = "^3.3.0"
psycopg2-binary = "^2.9.5"
pydantic-settings = "^2.0.3"
orjson = "^3.9.9"
aiocache = "^0.12.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
mypy = "^1.6.0"
uvicorn = "^0.23.2"
types-python-jose = "^3.3.4.8"
types-passlib = "^1.7.7.13"
black = "^23.10.1"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
httpx = "^0.25.0"

[tool.mypy]
exclude = ["app/tests"]
strict = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = "aiocache.*"
ignore_missing_imports = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
